Update New-UnityLUN.ps1
In our environment we need for a LUN to be visible to the Hosts with the same HLU. 
I've added a variable in which to record the desired hluID added the added it to the $blockHostAccessParam and this way I solved the problem for myself. 
It easily could be reworked to generate sequential HLU IDs if preferred. 
I don't know if it'll be of any use to others, but decided to share.